1.19.0
updated the componet/cart style by adding beter icon, and css positioning, added front end for cart to have a form to display if there is not shipping address on the logged in users user model, updated the query so that it returns address info. updated style and positioning of form.