1.34.0
Change logic so that similar order mapping in admin and user dash are one mapping component that take a prop of the data export. Updated stripe checkout resolver to correctly impliment a tax rate. set an agree token and created a new util/redirector class object which checks the validity of the users tokens, then logs out if Atoken is expired and redirects to landing page if null or expired agree token, added isIdsub and isIdRej properties to typedefs and user model in preperation for better prompting to the user upon completing verification process.