1.35.0
Modified logic to enhance the verification process, displaying whether submissions have been accepted or rejected dynamically. Additionally, optimized queries by reducing from 2 to 1. The query is now performed in the parent function component, and the data is passed down to the child components. Change 2: Maintained Link tags instead of <a> tags, especially with image buckets, to boost loading speed. However, faced an issue where switching pages wouldnt scroll to the top. To address this, a helper function was created using useEffect to enable smooth scrolling to the top of the page. Change 3: Optimized the security logic on the success page, ensuring a safer user experience.