1.47.0
Updated logic to include quantity field for purchased orders, handling purchase quantity through mapping and looping based on the purchaseQuantity field. Created a new map to prevent duplicate pushes and increase quantity, enabling dynamic rendering of products and quantities.

In addition, introduced a modal for updating products and outlined a resolver. Updated the query to ensure that cloverId is a valid property of the Product model.